Add an example about using IsometricStarPolygon class to the demo
