+++ demo/demo.html
@@ -63,11 +63,14 @@
     <h2>IsometricPath using curves</h2>
     <div id="demo6" class="demo"></div>
 
-    <h2>Dragging interaction with boundaries</h2>
+    <h2>Drawing Star Polygons</h2>
     <div id="demo7" class="demo"></div>
 
-    <h2>Using textures</h2>
+    <h2>Dragging interaction with boundaries</h2>
     <div id="demo8" class="demo"></div>
+
+    <h2>Using textures</h2>
+    <div id="demo9" class="demo"></div>
 
     <script async defer src="https://buttons.github.io/buttons.js"></script>
   </body>
