+++ docs/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><meta name="viewport" content="width=device-width,user-scalable=yes"/><title>Isometric demo</title><meta name="description" content="A JavaScript library written in TypeScript to create isometric projections"/><link rel="icon" href="favicon.png"><script defer="defer" src="scripts/bundle.js"></script><link href="styles/styles.css" rel="stylesheet"></head><body><header class="separator"><a href="https://github.com/elchininet/isometric" target="_blank" rel="noopener noreferrer"><img class="logo" src="images/logo.svg"/></a></header><div class="github-buttons"><a class="github-button" href="https://github.com/elchininet/isometric/archive/master.zip" data-icon="octicon-cloud-download" data-size="large" aria-label="Download elchininet/isometric on GitHub">Download</a> <a class="github-button" href="https://github.com/elchininet/isometric/fork" data-icon="octicon-repo-forked" data-size="large" aria-label="Fork elchininet/isometric on GitHub">Fork</a> <a class="github-button" href="https://github.com/elchininet/isometric" data-icon="octicon-star" data-size="large" aria-label="Star elchininet/isometric on GitHub">Star</a></div><h2>Basic example with interaction</h2><div id="demo1" class="demo"></div><h2>Basic example with interaction and animations</h2><div id="demo2" class="demo"></div><h2>Basic example with interaction using text</h2><div id="demo3" class="demo"></div><h2>IsometricPath using drawing methods with interaction</h2><div id="demo4" class="demo"></div><h2>IsometricPath using drawing commands</h2><div id="demo5" class="demo"></div><h2>IsometricPath using curves</h2><div id="demo6" class="demo"></div><h2>Dragging interaction with boundaries</h2><div id="demo7" class="demo"></div><h2>Using textures</h2><div id="demo8" class="demo"></div><script async defer="defer" src="https://buttons.github.io/buttons.js"></script></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><meta name="viewport" content="width=device-width,user-scalable=yes"/><title>Isometric demo</title><meta name="description" content="A JavaScript library written in TypeScript to create isometric projections"/><link rel="icon" href="favicon.png"><script defer="defer" src="scripts/bundle.js"></script><link href="styles/styles.css" rel="stylesheet"></head><body><header class="separator"><a href="https://github.com/elchininet/isometric" target="_blank" rel="noopener noreferrer"><img class="logo" src="images/logo.svg"/></a></header><div class="github-buttons"><a class="github-button" href="https://github.com/elchininet/isometric/archive/master.zip" data-icon="octicon-cloud-download" data-size="large" aria-label="Download elchininet/isometric on GitHub">Download</a> <a class="github-button" href="https://github.com/elchininet/isometric/fork" data-icon="octicon-repo-forked" data-size="large" aria-label="Fork elchininet/isometric on GitHub">Fork</a> <a class="github-button" href="https://github.com/elchininet/isometric" data-icon="octicon-star" data-size="large" aria-label="Star elchininet/isometric on GitHub">Star</a></div><h2>Basic example with interaction</h2><div id="demo1" class="demo"></div><h2>Basic example with interaction and animations</h2><div id="demo2" class="demo"></div><h2>Basic example with interaction using text</h2><div id="demo3" class="demo"></div><h2>IsometricPath using drawing methods with interaction</h2><div id="demo4" class="demo"></div><h2>IsometricPath using drawing commands</h2><div id="demo5" class="demo"></div><h2>IsometricPath using curves</h2><div id="demo6" class="demo"></div><h2>Drawing Star Polygons</h2><div id="demo7" class="demo"></div><h2>Dragging interaction with boundaries</h2><div id="demo8" class="demo"></div><h2>Using textures</h2><div id="demo9" class="demo"></div><script async defer="defer" src="https://buttons.github.io/buttons.js"></script></body></html>
